RDKCOM-2762: REFPLTV-1146 Bring up 64bit support in Rpi4 (for both kernel an...
Reason for change: to match datatype compatibilty
Test Procedure: Build should be successful. tr69hostif should build fine.
Risks: Low/None.
(cherry picked from commit 10a38ce23dbc2f73cb2fd97d4f1d7039f45c3ae0)

+++ src/hostif/handlers/src/hostIf_dsClient_ReqHandler.cpp
@@ -449,7 +449,7 @@ void DSClientReqHandler::checkForUpdates()
     GList *devList = hostIf_STBServiceHDMI::getAllInstances();
     for(elem = devList; elem; elem = elem->next,index++)
     {
-        hostIf_STBServiceHDMI *pIface = hostIf_STBServiceHDMI::getInstance((int)elem->data);
+        hostIf_STBServiceHDMI *pIface = hostIf_STBServiceHDMI::getInstance((intptr_t)elem->data);
         if(pIface)
         {
 			pIface->doUpdates(mUpdateCallback);
@@ -461,7 +461,7 @@ void DSClientReqHandler::checkForUpdates()
 	devList = hostIf_STBServiceVideoOutput::getAllInstances();
     for(elem = devList; elem; elem = elem->next,index++)
     {
-        hostIf_STBServiceVideoOutput *pIface = hostIf_STBServiceVideoOutput::getInstance((int)elem->data);
+        hostIf_STBServiceVideoOutput *pIface = hostIf_STBServiceVideoOutput::getInstance((intptr_t)elem->data);
         if(pIface)
         {
 			pIface->doUpdates(mUpdateCallback);
@@ -473,7 +473,7 @@ void DSClientReqHandler::checkForUpdates()
 	devList = hostIf_STBServiceVideoDecoder::getAllInstances();
     for(elem = devList; elem; elem = elem->next,index++)
     {
-        hostIf_STBServiceVideoDecoder *pIface = hostIf_STBServiceVideoDecoder::getInstance((int)elem->data);
+        hostIf_STBServiceVideoDecoder *pIface = hostIf_STBServiceVideoDecoder::getInstance((intptr_t)elem->data);
         if(pIface)
         {
 			pIface->doUpdates(mUpdateCallback);
@@ -486,7 +486,7 @@ void DSClientReqHandler::checkForUpdates()
 
     for(elem = devList; elem; elem = elem->next,index++)
     {
-        hostIf_STBServiceAudioInterface *pIfaceAudio = hostIf_STBServiceAudioInterface::getInstance((int)elem->data);
+        hostIf_STBServiceAudioInterface *pIfaceAudio = hostIf_STBServiceAudioInterface::getInstance((intptr_t)elem->data);
 
         if(pIfaceAudio)
         {
@@ -499,7 +499,7 @@ void DSClientReqHandler::checkForUpdates()
     
     for (elem = devList; elem; elem = elem->next, index++)
     {
-        hostIf_STBServiceSPDIF *pIfaceSPDIF = hostIf_STBServiceSPDIF::getInstance((int)elem->data);
+        hostIf_STBServiceSPDIF *pIfaceSPDIF = hostIf_STBServiceSPDIF::getInstance((intptr_t)elem->data);
         if (pIfaceSPDIF)
         {
             pIfaceSPDIF->doUpdates(mUpdateCallback);
